Enable baremetal post operations
Added these default parameters:
- baremetal_instackenv;
- baremetal_network_isolation;
by default valued to point files in the stack home directory.

Added to the main task the push of (only when undercloud_setup is set to
"baremetal"):
- Baremetal instackenv.json file;
- Baremetal network-environment.yaml file;
- Baremetal nic-configs directory (if passed);
All these must be passed by the user since are depending from the
environment.

Added to the undercloud-install-post.sh bash script (only when
undercloud_setup is set to "baremetal") these actions:
- External network device creation associated with the
  ExternalNetworkVlanID dynamically read from network-environment.yaml;
- Association between the discovered baremetal nodes and the baremetal
  profile;

Change-Id: I085deb448eb81276f033539ad9c68f0b53af9667